requirements.txt
colorzero==2.0
customtkinter==5.2.2
darkdetect==0.8.0
filelock==3.19.1
fsspec==2025.9.0
gpiozero==2.0.1
Jinja2==3.1.6
# lgpio==0.2.2.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pigpio==1.78
pillow==11.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
# RPi.GPIO==0.7.1
six==1.17.0
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
tzdata==2025.2
